mba: moving the boot event log parsing to the MBA subdirectory
