+++ keylime/da/attest.py
@@ -41,6 +41,7 @@ def main() -> None:
     args = parser.parse_args()
 
     mba.load_policy_engine()
+    mba.load_parser_engine()
 
     rmcb = config.get("registrar", "durable_attestation_import", fallback="")
     rmc = record.get_record_mgt_class(rmcb)
