+++ keylime/mba/elchecking/__main__.py
@@ -2,7 +2,7 @@ import argparse
 import json
 import sys
 
-from keylime.tpm import tpm_main
+from keylime.mba.elparsing import tpm2_tools_elparser
 
 from . import policies
 
@@ -28,8 +28,7 @@ if policy is None:
 refstate_str = args.refstate_file.read()
 refstate = json.loads(refstate_str)
 log_bin = args.eventlog_file.read()
-Tpm = tpm_main.Tpm()
-_, log_data = Tpm.parse_binary_bootlog(log_bin)
+_, log_data = tpm2_tools_elparser.parse_binary_bootlog(log_bin)
 with open("/tmp/parsed.json", "wt", encoding="utf-8") as log_data_file:
     log_data_file.write(json.dumps(log_data, indent=True))
 why_not = policy.evaluate(refstate, log_data)
