+++ keylime/mba/elparsing/__init__.py
@@ -1,0 +1,3 @@
+# Everything in this package is moved here from other parts of the keylime code.
+# This code is in charge of running `tpm2_eventlog` (part of the Intel tpm2 tool chain)
+# to parse binary boot logs into JSON.
